+++ src/crawlee/__init__.py
@@ -1,5 +1,7 @@
 from importlib import metadata
 
+from ._autoscaling.types import ConcurrencySettings
+from ._models import Request
 from ._utils.globs import Glob
 
 __version__ = metadata.version('crawlee')
