+++ src/crawlee/_autoscaling/autoscaled_pool.py
@@ -9,8 +9,8 @@ from datetime import timedelta
 from logging import getLogger
 from typing import TYPE_CHECKING, Awaitable, Callable
 
+from crawlee._autoscaling.types import ConcurrencySettings
 from crawlee._utils.recurring_task import RecurringTask
-from crawlee.types import ConcurrencySettings
 
 if TYPE_CHECKING:
     from crawlee._autoscaling import SystemStatus
