+++ src/crawlee/_memory_storage_client/request_queue_client.py
@@ -16,16 +16,7 @@ from crawlee._memory_storage_client._creation_management import (
     find_or_create_client_by_id_or_name_inner,
     persist_metadata_if_enabled,
 )
-from crawlee._utils.crypto import crypto_random_object_id
-from crawlee._utils.data_processing import (
-    filter_out_none_values_recursively,
-    raise_on_duplicate_storage,
-    raise_on_non_existing_storage,
-)
-from crawlee._utils.file import force_remove, force_rename, json_dumps
-from crawlee._utils.requests import unique_key_to_request_id
-from crawlee.base_storage_client import BaseRequestQueueClient
-from crawlee.models import (
+from crawlee._models import (
     BatchRequestsOperationResponse,
     ProcessedRequest,
     ProlongRequestLockResponse,
@@ -35,7 +26,16 @@ from crawlee.models import (
     RequestQueueMetadata,
     UnprocessedRequest,
 )
-from crawlee.types import StorageTypes
+from crawlee._types import StorageTypes
+from crawlee._utils.crypto import crypto_random_object_id
+from crawlee._utils.data_processing import (
+    filter_out_none_values_recursively,
+    raise_on_duplicate_storage,
+    raise_on_non_existing_storage,
+)
+from crawlee._utils.file import force_remove, force_rename, json_dumps
+from crawlee._utils.requests import unique_key_to_request_id
+from crawlee.base_storage_client import BaseRequestQueueClient
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
