+++ src/crawlee/base_storage_client/_base_dataset_client.py
@@ -6,8 +6,8 @@ from typing import TYPE_CHECKING, AsyncContextManager, AsyncIterator
 if TYPE_CHECKING:
     from httpx import Response
 
-    from crawlee.models import DatasetItemsListPage, DatasetMetadata
-    from crawlee.types import JsonSerializable
+    from crawlee._models import DatasetItemsListPage, DatasetMetadata
+    from crawlee._types import JsonSerializable
 
 
 class BaseDatasetClient(ABC):
