+++ src/crawlee/base_storage_client/_base_request_queue_collection_client.py
@@ -4,7 +4,7 @@ from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from crawlee.models import RequestQueueListPage, RequestQueueMetadata
+    from crawlee._models import RequestQueueListPage, RequestQueueMetadata
 
 
 class BaseRequestQueueCollectionClient(ABC):
