+++ src/crawlee/basic_crawler/_basic_crawler.py
@@ -23,7 +23,9 @@ from crawlee._autoscaling import AutoscaledPool
 from crawlee._autoscaling.snapshotter import Snapshotter
 from crawlee._autoscaling.system_status import SystemStatus
 from crawlee._log_config import CrawleeLogFormatter
+from crawlee._models import BaseRequestData, DatasetItemsListPage, Request, RequestState
 from crawlee._statistics import Statistics
+from crawlee._types import BasicCrawlingContext, HttpHeaders, RequestHandlerRunResult, SendRequestFunction
 from crawlee._utils.urls import convert_to_absolute_url, is_url_absolute
 from crawlee._utils.wait import wait_for
 from crawlee.basic_crawler._context_pipeline import ContextPipeline
@@ -36,22 +38,16 @@ from crawlee.errors import (
     UserDefinedErrorHandlerError,
 )
 from crawlee.http_clients import HttpxHttpClient
-from crawlee.models import BaseRequestData, DatasetItemsListPage, Request, RequestState
 from crawlee.router import Router
 from crawlee.sessions import SessionPool
 from crawlee.storages import Dataset, KeyValueStore, RequestQueue
-from crawlee.types import (
-    BasicCrawlingContext,
-    ConcurrencySettings,
-    HttpHeaders,
-    RequestHandlerRunResult,
-    SendRequestFunction,
-)
 
 if TYPE_CHECKING:
     import re
 
+    from crawlee._autoscaling.types import ConcurrencySettings
     from crawlee._statistics import FinalStatistics, StatisticsState
+    from crawlee._types import HttpMethod, JsonSerializable
     from crawlee.configuration import Configuration
     from crawlee.events._event_manager import EventManager
     from crawlee.http_clients import BaseHttpClient, HttpResponse
@@ -59,7 +55,6 @@ if TYPE_CHECKING:
     from crawlee.sessions import Session
     from crawlee.storages._dataset import GetDataKwargs, PushDataKwargs
     from crawlee.storages._request_provider import RequestProvider
-    from crawlee.types import HttpMethod, JsonSerializable
 
 TCrawlingContext = TypeVar('TCrawlingContext', bound=BasicCrawlingContext, default=BasicCrawlingContext)
 ErrorHandler = Callable[[TCrawlingContext, Exception], Awaitable[Union[Request, None]]]
