+++ src/crawlee/beautifulsoup_crawler/_beautifulsoup_crawler.py
@@ -8,6 +8,7 @@ from bs4 import BeautifulSoup, Tag
 from pydantic import ValidationError
 from typing_extensions import Unpack
 
+from crawlee._models import BaseRequestData
 from crawlee._utils.blocked import RETRY_CSS_SELECTORS
 from crawlee._utils.urls import convert_to_absolute_url, is_url_absolute
 from crawlee.basic_crawler import BasicCrawler, BasicCrawlerOptions, ContextPipeline
@@ -16,10 +17,9 @@ from crawlee.enqueue_strategy import EnqueueStrategy
 from crawlee.errors import SessionError
 from crawlee.http_clients import HttpxHttpClient
 from crawlee.http_crawler import HttpCrawlingContext
-from crawlee.models import BaseRequestData
 
 if TYPE_CHECKING:
-    from crawlee.types import AddRequestsKwargs, BasicCrawlingContext
+    from crawlee._types import AddRequestsKwargs, BasicCrawlingContext
 
 
 class BeautifulSoupCrawler(BasicCrawler[BeautifulSoupCrawlingContext]):
