+++ src/crawlee/http_clients/_base.py
@@ -5,11 +5,11 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING, Protocol
 
 if TYPE_CHECKING:
+    from crawlee._models import Request
     from crawlee._statistics import Statistics
-    from crawlee.models import Request
+    from crawlee._types import HttpHeaders, HttpMethod
     from crawlee.proxy_configuration import ProxyInfo
     from crawlee.sessions import Session
-    from crawlee.types import HttpHeaders, HttpMethod
 
 
 class HttpResponse(Protocol):
