+++ src/crawlee/http_clients/curl_impersonate.py
@@ -22,11 +22,11 @@ if TYPE_CHECKING:
 
     from curl_cffi.requests import Response
 
+    from crawlee._models import Request
     from crawlee._statistics import Statistics
-    from crawlee.models import Request
+    from crawlee._types import HttpHeaders, HttpMethod
     from crawlee.proxy_configuration import ProxyInfo
     from crawlee.sessions import Session
-    from crawlee.types import HttpHeaders, HttpMethod
 
 
 class _CurlImpersonateResponse:
