+++ src/crawlee/playwright_crawler/_playwright_crawler.py
@@ -6,20 +6,20 @@ from typing import TYPE_CHECKING, Literal
 from pydantic import ValidationError
 from typing_extensions import Unpack
 
+from crawlee._models import BaseRequestData
 from crawlee._utils.blocked import RETRY_CSS_SELECTORS
 from crawlee._utils.urls import convert_to_absolute_url, is_url_absolute
 from crawlee.basic_crawler import BasicCrawler, BasicCrawlerOptions, ContextPipeline
 from crawlee.browsers import BrowserPool
 from crawlee.enqueue_strategy import EnqueueStrategy
 from crawlee.errors import SessionError
-from crawlee.models import BaseRequestData
 from crawlee.playwright_crawler._playwright_crawling_context import PlaywrightCrawlingContext
 from crawlee.playwright_crawler._utils import infinite_scroll
 
 if TYPE_CHECKING:
     from collections.abc import AsyncGenerator
 
-    from crawlee.types import AddRequestsKwargs, BasicCrawlingContext
+    from crawlee._types import AddRequestsKwargs, BasicCrawlingContext
 
 
 class PlaywrightCrawler(BasicCrawler[PlaywrightCrawlingContext]):
