+++ src/crawlee/storages/_request_list.py
@@ -11,7 +11,7 @@ from crawlee.storages._request_provider import RequestProvider
 if TYPE_CHECKING:
     from collections.abc import Sequence
 
-    from crawlee.models import BaseRequestData, Request
+    from crawlee._models import BaseRequestData, Request
 
 
 class RequestList(RequestProvider):
