+++ tests/unit/basic_crawler/test_context_pipeline.py
@@ -7,11 +7,11 @@ from unittest.mock import AsyncMock
 
 import pytest
 
+from crawlee import Request
+from crawlee._types import BasicCrawlingContext
 from crawlee.basic_crawler import ContextPipeline
 from crawlee.errors import ContextPipelineFinalizationError, ContextPipelineInitializationError, RequestHandlerError
-from crawlee.models import Request
 from crawlee.sessions._session import Session
-from crawlee.types import BasicCrawlingContext
 
 
 @dataclass(frozen=True)
