+++ tests/unit/http_clients/test_curl_impersonate.py
@@ -5,10 +5,10 @@ from typing import TYPE_CHECKING
 
 import pytest
 
+from crawlee import Request
 from crawlee._statistics import Statistics
 from crawlee.errors import ProxyError
 from crawlee.http_clients.curl_impersonate import CurlImpersonateHttpClient
-from crawlee.models import Request
 
 if TYPE_CHECKING:
     from crawlee.proxy_configuration import ProxyInfo
