+++ tests/unit/proxy_configuration/test_tiers.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from crawlee.models import Request
+from crawlee import Request
 from crawlee.proxy_configuration import ProxyConfiguration
